backend/requirements.txt
alembic==1.14.1
amqp==5.3.1
appdirs==1.4.3
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.8.1
astroid==2.3.3
asttokens==3.0.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==19.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
billiard==4.2.1
black==19.3b0
blinker==1.4
blis==0.7.11
Brlapi==0.7.0
catalogue==2.0.10
celery==5.5.3
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmdtest==0.32+git
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
confection==0.1.5
configobj==5.0.6
contourpy==1.1.1
crudini==0.9.3
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.11
dbus-python==1.2.16
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
distlib==0.3.9
distro==1.4.0
django==4.2.24
django-cors-headers==4.4.0
djangorestframework==3.15.2
entrypoints==0.3
et-xmlfile==2.0.0
executing==2.2.0
filelock==3.16.1
flake8==3.7.9
fonttools==4.57.0
greenlet==3.1.1
grpcio==1.16.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
IMDbPY==6.8
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniparse==0.4
ipykernel==6.29.5
ipython==8.12.3
isort==4.3.4
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.8.1
keyring==18.0.1
kiwisolver==1.4.7
kombu==5.5.4
langcodes==3.4.1
language-data==1.3.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
logilab-common==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
marisa-trie==1.2.1
Markdown==3.1.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mccabe==0.6.1
murmurhash==1.0.13
mysql-connector==2.2.9
nemo-emblems==5.2.0
nest-asyncio==1.6.0
netaddr==0.7.19
netifaces==0.10.4
ngrok==1.4.0
numpy==1.24.4
oauthlib==3.1.0
onboard==1.4.1
openpyxl==3.1.5
packaging==20.3
PAM==0.4.2
panda==0.3.1
pandas==2.0.3
parso==0.8.4
pathlib-abc==0.1.1
pathy==0.11.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==4.3.6
preshed==3.0.10
prompt-toolkit==3.0.51
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.9.10
pure-eval==0.2.3
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pydantic==1.10.22
pyflakes==2.1.1
pygments==2.19.2
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
pytz==2025.2
pyxdg==0.26
PyYAML==5.3.1
pyzmq==27.0.0
redis==6.1.1
reportlab==3.5.34
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
smart-open==6.4.0
soupsieve==1.9.5
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.41
sqlparse==0.5.3
srsly==2.4.8
stack-data==0.6.3
systemd-python==234
thinc==8.1.12
tinycss2==1.0.2
tldextract==2.2.1
toml==0.10.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
ttystatus==0.38
typer==0.9.4
typing-extensions==4.13.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
vine==5.1.0
virtualenv==20.31.2
wadllib==1.3.3
wasabi==1.1.3
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.11.2
xkit==0.0.0
xlrd==1.1.0
youtube-dl==2021.4.26
zipp==3.20.2
